+++ internal/model/user.go
@@ -12,7 +12,7 @@ type Users struct {
 	Password    string    `json:"password,omitempty"`
 	PhoneNumber string    `json:"phone_number"`
 	Address     string    `json:"address"`
-	Age         int8      `json:"age"`
+	Age         int       `json:"age"`
 	Role        Role      `json:"role"` //Enums Role
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
